requirements.txt
asgiref==3.6.0
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
pytz==2022.7
sqlparse==0.4.3
